venv integration

+++ packages/mate/mate_project/python.py
@@ -1,44 +1,69 @@
 import os
 import subprocess
+import sys
+import shutil
 
 
 class Package:
     def __init__(self, name: str, version: str):
         self.name = name
-        self.version = version
+        self.version = version  # TODO: add version check
 
     def __hash__(self):
         return self.name.__hash__()
 
-    def __repr__(self):
+    def __repr__(self):  #
         return f"Package({self.name}, {self.version})"
+
+    def __eq__(self, other):
+        if isinstance(other, str):
+            return self.name == other
+        elif isinstance(other, Package):
+            return self.name == other.name
+        else:
+            return False
 
 
 class Python:
-    def __init__(self, mate_dir: str) -> None:
-
+    def __init__(self, mate_dir: str, venv: bool = True) -> None:
         self.mate_dir = mate_dir
         self.venv_path = os.path.join(self.mate_dir, "venv")
         os.makedirs(self.venv_path, exist_ok=True)
-        self.python_path = os.path.join(self.venv_path, "bin", "python")
-        if not os.path.exists(self.python_path):
+        self.python_path = (
+            os.path.join(self.venv_path, "bin", "python") if venv else sys.executable
+        )
+        if venv and not os.path.exists(self.python_path):
             print(f"Creating virtual environment in {self.venv_path}")
             os.system(f"python -m venv {self.venv_path}")
             os.system(f"{self.python_path} -m pip install --upgrade pip")
+            os.system(f"{self.python_path} -m pip install wheel")
             os.system(f"{self.python_path} -m pip install pipreqs")
             self.__install_mate()
-
-        with open(os.path.join(self.venv_path, "pyvenv.cfg")) as f:
+        if venv:
+            with open(os.path.join(self.venv_path, "pyvenv.cfg")) as f:
+                self.cfg = {
+                    key.strip(): val.strip()
+                    for key, val in [line.split("=") for line in f.readlines()]
+                }
+        else:
             self.cfg = {
-                key.strip(): val.strip()
-                for key, val in [line.split("=") for line in f.readlines()]
+                "home": sys.executable,
+                "include-system-site-packages": "true",
+                "version": sys.version,
             }
+
         self.pipreqs_path = os.path.join(self.venv_path, "bin", "pipreqs")
-        if not os.path.exists(self.pipreqs_path):
+        self.requirements_path = os.path.join(self.mate_dir, "requirements.txt")
+        command = f"{self.python_path} -m pip freeze"
+        output = subprocess.check_output(command, shell=True).decode("utf-8")
+        with open(self.requirements_path, "w") as f:
+            f.write(output)
+
+        self.installed_packages = self.__requirements_to_packages(
+            self.requirements_path
+        )
+        if not "pipreqs" in self.installed_packages:
             self.pip_install("pipreqs")
-        self.requirements_path = os.path.join(self.venv_path, "requirements.txt")
-        os.system(f"{self.python_path} -m pip freeze > {self.requirements_path}")
-        # checks the output of a subprocess using the check_output function
         try:
             # FIXME: handle different mate versions
             command = f"{self.python_path} -m mate"
@@ -51,7 +76,6 @@ class Python:
             self.__install_mate()
 
     def __install_mate(self):
-
         mate_location = os.sep.join(__file__.split(os.sep)[:-4])
         os.system(f"{self.python_path} -m pip install -e {mate_location}")
 
@@ -60,10 +84,6 @@ class Python:
             reqs = [l.split("==") for l in f.readlines()]
         return tuple(Package(r[0], r[1]) for r in reqs if len(r) > 1)
 
-    # def venv(self, command: str):
-    #     os.system(f"source {os.path.join(self.venv_path, 'bin', 'activate')}")
-    #    #self.
-
     def pip_install(self, *packages: str):
         self(f'-m pip install {" ".join(packages)}')
 
@@ -71,12 +91,8 @@ class Python:
         os.system(f"{self.python_path} -m pip {command}")
 
     @property
-    def installed_packages(self):
-        return self.__requirements_to_packages(self.requirements_path)
-
-    @property
     def version(self):
-        return self.cfg["version"]
+        return self.cfg["version"].split()[0]
 
     def __call__(self, command: str):
         os.system(f"{self.python_path} {command}")
